requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-mail>=0.10.0
flask-login>=0.6.0
flask-wtf>=1.2.0
wtforms>=3.1.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
stripe>=8.0.0
apscheduler>=3.10.0
reportlab>=4.0.0
python-dotenv>=1.0.0
email-validator>=2.1.0
